fix: mudança no banco para receber ordersItems

+++ src/main/java/com/unyx/ticketeira/dto/batch/BatchCreateRequest.java
@@ -1,4 +1,6 @@
 package com.unyx.ticketeira.dto.batch;
 
-public record BatchCreateRequest(String name, int quantity, Double price, boolean requiresIdentification) {
+import java.math.BigDecimal;
+
+public record BatchCreateRequest(String name, int quantity, BigDecimal price, boolean requiresIdentification) {
 }
